eRPC updates 12/2019
-- improved the test_callbacks logic to be more understandable and to allow requested callback execution on the server side
-- transportArbitrator::prepareClientReceive modified to avoid incorrect return value type
-- the ClientManager and the ArbitratedClientManager updated to avoid performing client requests when the previous serialization phase fails
-- generate the shim code for destroy of statically allocated services
-- documentation update for v1.7.3

+++ erpc_c/setup/erpc_setup_mbf_rpmsg.cpp
@@ -47,7 +47,7 @@ public:
     virtual MessageBuffer create(void)
     {
         void *buf = NULL;
-        unsigned long size = 0;
+        uint32_t size = 0;
         buf = rpmsg_lite_alloc_tx_buffer(m_rpmsg, &size, RL_BLOCK);
 
         assert(NULL != buf);
@@ -65,7 +65,7 @@ public:
         void *tmp = (void *)buf->get();
         if (tmp)
         {
-            int ret;
+            int32_t ret;
             ret = rpmsg_lite_release_rx_buffer(m_rpmsg, tmp);
             if (ret != RL_SUCCESS)
             {
