+++ erpc_c/setup/erpc_setup_rpmsg_lite_remote.cpp
@@ -23,8 +23,8 @@ static ManuallyConstructed<RPMsgTransport> s_transport;
 // Code
 ////////////////////////////////////////////////////////////////////////////////
 
-erpc_transport_t erpc_transport_rpmsg_lite_remote_init(unsigned long src_addr, unsigned long dst_addr,
-                                                       void *start_address, int rpmsg_link_id, rpmsg_ready_cb ready,
+erpc_transport_t erpc_transport_rpmsg_lite_remote_init(uint32_t src_addr, uint32_t dst_addr, void *start_address,
+                                                       int32_t rpmsg_link_id, rpmsg_ready_cb ready,
                                                        char *nameservice_name)
 {
     s_transport.construct();
