+++ erpc_c/setup/erpc_setup_rpmsg_lite_rtos_master.cpp
@@ -37,8 +37,7 @@ static ManuallyConstructed<RPMsgRTOSTransport> s_transport;
 // Code
 ////////////////////////////////////////////////////////////////////////////////
 
-erpc_transport_t erpc_transport_rpmsg_lite_rtos_master_init(unsigned long src_addr, unsigned long dst_addr,
-                                                            int rpmsg_link_id)
+erpc_transport_t erpc_transport_rpmsg_lite_rtos_master_init(uint32_t src_addr, uint32_t dst_addr, int32_t rpmsg_link_id)
 {
     s_transport.construct();
     if (s_transport->init(src_addr, dst_addr, rpmsg_lite_base, SH_MEM_TOTAL_SIZE, rpmsg_link_id) == kErpcStatus_Success)
