+++ erpc_c/transports/erpc_rpmsg_tty_rtos_transport.h
@@ -67,8 +67,8 @@ public:
      * @retval kErpcStatus_InitFailed When rpmsg init function wasn't executed
      * successfully.
      */
-    virtual erpc_status_t init(unsigned long src_addr, unsigned long dst_addr, void *base_address, unsigned long length,
-                               int rpmsg_link_id);
+    virtual erpc_status_t init(uint32_t src_addr, uint32_t dst_addr, void *base_address, uint32_t length,
+                               int32_t rpmsg_link_id);
 
     /*!
      * @brief This function call RPMsg rtos init function - as RPMsg remote
@@ -88,7 +88,7 @@ public:
      * @retval kErpcStatus_InitFailed When rpmsg init function wasn't executed
      * successfully.
      */
-    virtual erpc_status_t init(unsigned long src_addr, unsigned long dst_addr, void *base_address, int rpmsg_link_id,
+    virtual erpc_status_t init(uint32_t src_addr, uint32_t dst_addr, void *base_address, int32_t rpmsg_link_id,
                                void (*ready_cb)(void), char *nameservice_name);
 
     /*!
@@ -122,7 +122,7 @@ public:
     virtual void setCrc16(Crc16 *crcImpl);
 
 protected:
-    unsigned long m_dst_addr;                /*!< Destination address used by rpmsg. */
+    uint32_t m_dst_addr;                     /*!< Destination address used by rpmsg. */
     rpmsg_queue_handle m_rpmsg_queue;        /*!< Handle of RPMsg queue. */
     struct rpmsg_lite_endpoint *m_rpmsg_ept; /*!< Pointer to RPMsg Lite Endpoint structure. */
     Crc16 *m_crcImpl;                        //!< CRC object.
